fixed test locations
