+++ tests/terraform/runner/resources/resource_negative_value_without_var/variables_unscoped.tf
