+++ tests/terraform/runner/resources/resource_value_without_var/main.tf
