+++ tests/terraform/runner/resources/resource_value_without_var/variables.tf
